+++ src/main/java/com/shy_polarbear/server/domain/auth/repository/redis/RedisRefreshTokenRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RedisRefreshTokenRepository extends CrudRepository<RefreshToken,String> {
+public interface RedisRefreshTokenRepository extends CrudRepository<RedisRefreshToken,String> {
 }
